Feat/furystack updates

+++ src/CommandInterpreter/CustomCommands/wifi.h
@@ -2,8 +2,8 @@
 
 #include <Arduino.h>
 
-extern int32_t lastSentRssi;
+extern int8_t lastSentRssi;
 
-String getWifiMessage(int32_t rssi);
+String getWifiMessage(int8_t rssi);
 void wifiEvents();
 void setupWifi();
